# swagger.yaml
openapi: 3.0.0
info:
  title: Aplicação Restaurante Tech
  description: Na nossa região, um grupo de restaurantes decidiu contratar estudantes
                para construir um sistema de gestão para seus estabelecimentos. Essa decisão
                foi motivada pelo alto custo de sistemas individuais, o que levou os
                restaurantes a se unirem para desenvolver um sistema único e compartilhado.
                Esse sistema permitirá que os clientes escolham restaurantes com base na
                comida oferecida, em vez de se basearem na qualidade do sistema de gestão.
  version: 1.0.0
servers:
  - url: http://<hostname>/api/v1
paths:
...
  /restaurant/create:
    post:
      summary: Criar Restaurante
      description: Este endpoint cria um especifico Restaurante.
      operationId: createRestaurant
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/createRestaurant"
      responses:
        '200':
          description: Sucesso retorna o dados do Restaurante criado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/restaurantResponse"
        '400':
          description: Erro ao enviar a requisição, não reconhecendo o body enviado 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error400"
                  
        '404':
          description: Erro ao enviar a requisição, não reconhecendo o endpoint acessado 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error404"
                  
        '500':
          description: Erro ao acessar o serviço, falha ao tentar acessar o serviço
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error500"
      security:
      - bearerAuth: []
  /restaurant/update:
    put:
      summary: Alterar Restaurant
      description: Este endpoint altera o especifico Restaurant.
      operationId: updateRestaurant
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/updateRestaurant"
      responses:
        '200':
          description: Sucesso retorna o dados do Restaurant alterado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/restaurantResponse"
        '400':
          description: Erro ao enviar a requisição, não reconhecendo o body enviado 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error400"
                  
        '404':
          description: Erro ao enviar a requisição, não reconhecendo o endpoint acessado 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error404"
                  
        '500':
          description: Erro ao acessar o serviço, falha ao tentar acessar o serviço
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error500"
      security:
      - bearerAuth: []
  /restaurant/search/{nomeRestaurante}:
    get:
      summary: Busca Restaurant
      description: Este endpoint busca um especifico Restaurante.
      operationId: getRestaurant
      parameters:
        - name: nomeRestaurante
          in: path
          description: restaurant a ser encontrado
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sucesso retorna o dados do Cardapio buscado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/restaurantResponse"
        '400':
          description: Erro ao enviar a requisição, não reconhecendo o body enviado 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error400"
                  
        '404':
          description: Erro ao enviar a requisição, não reconhecendo o endpoint acessado 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error404"
                  
        '500':
          description: Erro ao acessar o serviço, falha ao tentar acessar o serviço
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error500"
      security:
      - bearerAuth: []
  /restaurant/list/number/{number}/page/{page}:
    get:
      summary: Lista os Restaurant registrados
      description: Este endpoint lista todos os Restaurants existente.
      operationId: listRestaurant
      parameters:
        - name: number
          in: path
          description: numero de pagina dividida na busca
          required: true
          schema:
            type: integer
            format: int32
        - name: page
          in: path
          description: numero de conteudo na pagina
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Sucesso retorna os dados buscado no banco com a data da criação, a data da ultima modificação, a regras com o Restaurant cadastrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/restaurantList"
        '400':
          description: Erro ao enviar a requisição, não reconhecendo o body enviado 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error400"
                  
        '404':
          description: Erro ao enviar a requisição, não reconhecendo o endpoint acessado 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error404"
                  
        '500':
          description: Erro ao acessar o serviço, falha ao tentar acessar o serviço
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error500" 
      security:
      - bearerAuth: []
  /restaurant/delete/{nomeRestaurante}:
    delete:
      summary: Desativa o Restaurant
      description: Este endpoint desativa o especifico Restaurant.
      operationId: deleteRestaurant
      parameters:
        - name: nomeRestaurante
          in: path
          description: restaurant a ser desativado
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sucesso retorna os dados da conta desativada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/restaurantResponse"
        '400':
          description: Erro ao enviar a requisição, não reconhecendo o body enviado 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error400"
                  
        '404':
          description: Erro ao enviar a requisição, não reconhecendo o endpoint acessado 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error404"
                  
        '500':
          description: Erro ao acessar o serviço, falha ao tentar acessar o serviço
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error500" 
components:
  schemas:
...
    restaurantList:
      type: object
      properties:
        number:
          type: integer
          description: numero de pagina
          example: 1
        page:
          type: integer
          description: quantidade na pagina
          example: 1
        total:
          type: integer
          description: total de pagina apresentado
          example: 1
        items:
          type: array
          description: itens buscados
          items:
            $ref: "#/components/schemas/restaurantResponse"
...
    createRestaurant:
      type: object
      properties:
        nomeRestaurante:
          type: string
          description: nome do restaurante
          example: Brasileiro
        tipo:
          type: string
          description: segmento do Restaurante
          example: prato tipico brasileiro
        endereco:
          type: object
          properties:
            CEP:
              type: string
              description: cep do endereço somente numeral
              example: 01001000
            logradouro:
              type: string
              description: endereço do presente no cep
              example: Rua Osvaldo
            bairro:
              type: string
              description: bairro do presente no cep
              example: Sé
            cidade:
              type: string
              description: cidade do presente no cep
              example: São Paulo
        abertura:
          type: string
          format: date-time
          description: data e hora que abre o restaurante
          example: 2025-10-11T23:26:29.0206602
        fechamento:
          type: string
          format: date-time
          description: data e hora que fecha o restaurante
          example: 2025-10-11T23:26:29.0206602
        responsavel:
          type: string
          description: o usuario responsavel pelo restaurante
          example: joao
    updateRestaurant:
      type: object
      properties:
        nomeRestaurante:
          type: string
          description: nome do restaurante
          example: Brasileiro
        tipo:
          type: string
          description: segmento do Restaurante
          example: prato tipico brasileiro
        endereco:
          type: object
          properties:
            CEP:
              type: string
              description: cep do endereço somente numeral
              example: 01001000
            logradouro:
              type: string
              description: endereço do presente no cep
              example: Rua Osvaldo
            bairro:
              type: string
              description: bairro do presente no cep
              example: Sé
            cidade:
              type: string
              description: cidade do presente no cep
              example: São Paulo
        abertura:
          type: string
          format: date-time
          description: data e hora que abre o restaurante
          example: 2025-10-11T23:26:29.0206602
        fechamento:
          type: string
          format: date-time
          description: data e hora que fecha o restaurante
          example: 2025-10-11T23:26:29.0206602
    restaurantResponse:
      type: object
      properties:
        nomeRestaurante:
          type: string
          description: nome do restaurante
          example: Brasileiro
        tipo:
          type: string
          description: segmento do Restaurante
          example: prato tipico brasileiro
        endereco:
          type: object
          properties:
            CEP:
              type: string
              description: cep do endereço somente numeral
              example: 01001000
            logradouro:
              type: string
              description: endereço do presente no cep
              example: Rua Osvaldo
            bairro:
              type: string
              description: bairro do presente no cep
              example: Sé
            cidade:
              type: string
              description: cidade do presente no cep
              example: São Paulo
        abertura:
          type: string
          format: date-time
          description: data e hora que abre o restaurante
          example: 2025-10-11T23:26:29.0206602
        fechamento:
          type: string
          format: date-time
          description: data e hora que fecha o restaurante
          example: 2025-10-11T23:26:29.0206602
        responsavel:
          type: string
          description: o usuario responsavel pelo restaurante
          example: joao
        actived:
          type: boolean
          description: indica se o restaurante esta disponivel
          example: true
        createAt:
          type: string
          format: date-time
          description: data de criação do restaurante
          example: 2025-10-11T23:26:29.0206602
        modifiedAt:
          type: string
          format: date-time
          description: data de modificação do restaurante
          example: 2025-12-11T23:26:29.0206602
    Error500:
      type: object
      properties:
        type:
          type: string
          description: informa o tipo de erro, com mensagem exceção
          example: java.lang.NullPointerException
        title:
          type: string
          description: informa o nome do erro, com mensagem status code
          example: Internal Server Error
        status:
          type: number
          description: Status Code do erro
          example: 500
        detail: 
          type: string
          description: descrição do erro
          example: Cannot invoke  because  is null
        instance: 
          type: string
          description: endpoint do erro
          example: /api/v1/
    Error404:
      type: object
      properties:
        type:
          type: string
          description: informa o tipo de erro, com mensagem exceção
          example: about:blank
        title:
          type: string
          description: informa o nome do erro, com mensagem status code
          example: Not Found
        status:
          type: number
          description: Status Code do erro
          example: 404
        detail: 
          type: string
          description: descrição do erro
          example: Serviço não encontrado
        instance: 
          type: string
          description: endpoint do erro
          example: /api/v1/
    Error400:
      type: object
      properties:
        type:
          type: string
          description: informa o tipo de erro, com mensagem exceção
          example: org.hibernate.sql.exec.ExecutionException
        title:
          type: string
          description: informa o nome do erro, com mensagem status code
          example: Bad Request
        status:
          type: number
          description: Status Code do erro
          example: 400
        detail: 
          type: string
          description: descrição do erro
          example: Erro na requisição
        instance: 
          type: string
          description: endpoint do erro
          example: /api/v1/
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT